+++ src/DependencyInjection/NexySlackExtension.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the Nexylan packages.
  *
@@ -12,34 +14,29 @@
 namespace Nexy\SlackBundle\DependencyInjection;
 
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\Alias;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
-use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
-class NexySlackExtension extends Extension
+final class NexySlackExtension extends Extension
 {
     /**
      * {@inheritdoc}
      */
-    public function load(array $configs, ContainerBuilder $container)
+    public function load(array $configs, ContainerBuilder $container): void
     {
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
 
-        $guzzleService = $config['guzzle_service'];
         $container->setParameter('nexy_slack.endpoint', $config['endpoint']);
+        $container->setAlias('nexy_slack.http.client', new Alias($config['http']['client'], false));
 
         // Unset the not needed keys for the Slack config.
-        unset($config['guzzle_service'], $config['endpoint']);
-
+        unset($config['http'], $config['endpoint']);
         $container->setParameter('nexy_slack.config', $config);
 
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('client.xml');
-
-        if ($guzzleService) {
-            $container->getDefinition('nexy_slack.client')->addArgument(new Reference($guzzleService));
-        }
     }
 }